Resolvendo erros no tratamento do parametro passado

+++ api/index.php
@@ -10,23 +10,23 @@ $id = filter_input(INPUT_GET, 'id', FILTER_VALIDATE_INT);
 $nome = filter_input(INPUT_GET, 'nome');
 $url = filter_input(INPUT_GET, 'url');
 $parametros = "video: $video, id: $id, nome:  $nome, url: $url";
-switch ($video) {
-    case "post":
-        addVideos($nome, $url);
-        print $parametros;
-        break;
-    case "put":
-        editVideo($id, $nome, $url);
-        print $parametros;
-        break;
-    case "delete":
-        deleteVideo($id);
-        print $parametros;
-        break;
-    default:
-        getVideos();
-        print $parametros;
-        break;
+
+if($video == "post") {
+    addVideos($nome, $url);
+    print $parametros;
 }
+elseif($video == "put") {
+    editVideo($id, $nome, $url);
+    print $parametros;
+}
+elseif($video == "delete") {
+    deleteVideo($id);
+    print $parametros;
+}
+else {
+    getVideos();
+    print $parametros;
+}
+
 
 ?>
